cilium: finally fix those fucking MTU issues
spent a lot of time trawling tailscale issue tracker, convinced it was
an issue with tailscale somehow.

turns out it was Cilium (it's always Cilium). I had already correctly
identified that Cilium needs help determining the MTU, it's just I got
it the wrong way around - it is autodetecting the 1280 of the tailscale0
interface, when it should be using the common 1500 of the main ethernet
link. zzzz....