synapse: eturnal (#87)
tried a little to get it working behind a loadbalancer with lots of
ports mapped, but the relaying wasn't working. I think I know why, and
maybe could have gotten it to work, but for this particularly unique
case it made more sense to just use host networking.

external-dns continues to be a solid tool. It supports this use case
admirably: all I had to do was put the Deployment pods behind a headless
service and add a special annotation to use the Node's external IP.

Interestingly, the docs made it sound like it was unavoidable that sub
records for each unique pod would be created. That does indeed happen
if the pod has a `.spec.hostname` set (as is the case with StatefulSets,
used in the doc example). If there's no hostname, it creates the top
level record only. Neato.