pixiecore: initial API mode deployment(#408)
the API is satisfied by a Bash script running under fcgiwrap and exposed
by nginx, in a sidecar container.

the script uses kubectl to lookup which Node is trying to boot, based on
a node label containing the MAC address. the script also decides whether
a boot is allowed, and which ignition configs to embed, based on
annotations on the Node.

the ignition configs that can be embedded are the ones that live in
ignition/ ... the pod decrypts these on boot with an age key that is
used only for this porpoise.

the script expects to run with hostPath access to a control-plane node,
so that it can generate join tokens and upload control-plane certs.