disable harbor container webhook
this might come as a shock, but operating harbor adds *more* complexity
to the cluster! specifically, the fucking thing seems to crash a lot,
and I can't be bothered figuring out why. for now just disable it. I
need to do another pass on the whole "works offline" theme anyway.